api/requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
celery[redis]
gunicorn
psycopg2
pydantic
